+++ app/src/main/kotlin/me/xizzhu/android/joshua/infra/arch/Interactor.kt
@@ -38,6 +38,26 @@ abstract class Interactor(dispatcher: CoroutineDispatcher) {
     }
 
     @UiThread
+    fun resume() {
+        onResume()
+    }
+
+    @CallSuper
+    @UiThread
+    protected open fun onResume() {
+    }
+
+    @UiThread
+    fun pause() {
+        onPause()
+    }
+
+    @CallSuper
+    @UiThread
+    protected open fun onPause() {
+    }
+
+    @UiThread
     fun stop() {
         onStop()
         coroutineScope.coroutineContext[Job]?.cancelChildren()