+++ docs/CHANGELOG.md
@@ -7,6 +7,9 @@ CHANGELOG
 - Should start / stop reading progress tracking on resume / pause
 - Supported hiding search button
 - Properly checked if root exception is JobCancellationException
+- Updated dependencies:
+  - Dagger to 2.25.3
+  - Mockito to 3.2.4
 
 #### v0.12.8 (2019-12-14)
 - Updated dependencies: